+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/topbar/TopBarUserField.kt
@@ -1,20 +1,25 @@
-@file:Suppress("HEADER_MISSING_IN_NON_SINGLE_CLASS_FILE")
+@file:Suppress("FILE_NAME_MATCH_CLASS")
 
 package com.saveourtool.save.frontend.components.topbar
 
 import com.saveourtool.save.domain.Role
 import com.saveourtool.save.frontend.components.modal.logoutModal
 import com.saveourtool.save.frontend.externals.fontawesome.*
+import com.saveourtool.save.info.UserInfo
 import com.saveourtool.save.v1
 import com.saveourtool.save.validation.FrontendRoutes
 
 import csstype.ClassName
-import react.FC
+import react.*
 import react.dom.aria.*
-import react.dom.html.ReactHTML
+import react.dom.html.ReactHTML.a
+import react.dom.html.ReactHTML.div
+import react.dom.html.ReactHTML.img
+import react.dom.html.ReactHTML.li
+import react.dom.html.ReactHTML.small
+import react.dom.html.ReactHTML.span
+import react.dom.html.ReactHTML.ul
 import react.router.useNavigate
-import react.useEffect
-import react.useState
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -24,10 +29,20 @@ import kotlinx.coroutines.isActive
 val topBarUserField = topBarUserField()
 
 /**
+ * [Props] of the top bar user field component
+ */
+external interface TopBarUserFieldProps : PropsWithChildren {
+    /**
+     * Currently logged in user or null
+     */
+    var userInfo: UserInfo?
+}
+
+/**
  * Displays the user's field
  */
 @Suppress("MAGIC_NUMBER", "LongMethod", "TOO_LONG_FUNCTION")
-private fun topBarUserField() = FC<TopBarProps> { props ->
+private fun topBarUserField() = FC<TopBarUserFieldProps> { props ->
     val (isLogoutModalOpen, setIsLogoutModalOpen) = useState(false)
     val (isAriaExpanded, setIsAriaExpanded) = useState(false)
     val scope = CoroutineScope(Dispatchers.Default)
@@ -39,20 +54,20 @@ private fun topBarUserField() = FC<TopBarProps> { props ->
             }
         }
     }
-    ReactHTML.ul {
+    ul {
         className = ClassName("navbar-nav ml-auto")
-        ReactHTML.div {
+        div {
             className = ClassName("topbar-divider d-none d-sm-block")
         }
         // Nav Item - User Information
-        ReactHTML.li {
+        li {
             className = ClassName("nav-item dropdown no-arrow")
             onClickCapture = {
                 setIsAriaExpanded {
                     !it
                 }
             }
-            ReactHTML.a {
+            a {
                 href = "#"
                 className = ClassName("nav-link dropdown-toggle")
                 id = "userDropdown"
@@ -61,24 +76,24 @@ private fun topBarUserField() = FC<TopBarProps> { props ->
                 ariaHasPopup = true.unsafeCast<AriaHasPopup>()
                 asDynamic()["data-toggle"] = "dropdown"
 
-                ReactHTML.div {
+                div {
                     className = ClassName("d-flex flex-row")
-                    ReactHTML.div {
+                    div {
                         className = ClassName("d-flex flex-column")
-                        ReactHTML.span {
+                        span {
                             className = ClassName("mr-2 d-none d-lg-inline text-gray-600")
                             +(props.userInfo?.name ?: "")
                         }
                         val globalRole = props.userInfo?.globalRole ?: Role.VIEWER
                         if (globalRole.isHigherOrEqualThan(Role.ADMIN)) {
-                            ReactHTML.small {
+                            small {
                                 className = ClassName("text-gray-400 text-justify")
                                 +globalRole.formattedName
                             }
                         }
                     }
                     props.userInfo?.avatar?.let {
-                        ReactHTML.img {
+                        img {
                             className =
                                     ClassName("ml-2 align-self-center avatar avatar-user width-full border color-bg-default rounded-circle fas mr-2")
                             src = "/api/$v1/avatar$it"
@@ -91,7 +106,7 @@ private fun topBarUserField() = FC<TopBarProps> { props ->
                 }
             }
             // Dropdown - User Information
-            ReactHTML.div {
+            div {
                 className = ClassName("dropdown-menu dropdown-menu-right shadow animated--grow-in${if (isAriaExpanded) " show" else ""}")
                 ariaLabelledBy = "userDropdown"
                 props.userInfo?.name?.let { name ->
